Copies chosen file/directory to selected directories
Copies chosen file/directory to selected directories. Multiple
directories can be selected to be backed up to.
Settings of which directories chosen can be saved and loaded.

Every time a file is saved to a directory, the original is overwritten
and  a backup is produced except for the first time saving in which the
file will only be copied over once without backup.